fix(dashboard): cut UI import chain from connection persist module (CI shard base-red)
tests/unit/agentrouter-connection-modal-fields.test.ts imports
connectionProviderSpecificData.ts, whose first import (providerPageHelpers)
reaches React icon-land; under the CI runner's Node 24 the @lobehub/icons
es/ build hits a CJS require -> 'Unexpected token export', failing Unit
shard 2/4 for EVERY PR. Extracted parseRoutingTagsInput/
parseExcludedModelsInput to a pure leaf (providerInputParsers.ts), kept
re-exports for UI importers, pointed the persist module at the leaf.
