+++ src/app/(dashboard)/dashboard/providers/[id]/components/modals/connectionProviderSpecificData.ts
@@ -1,4 +1,4 @@
-import { parseExcludedModelsInput, parseRoutingTagsInput } from "../../providerPageHelpers";
+import { parseExcludedModelsInput, parseRoutingTagsInput } from "../../providerInputParsers";
 import {
   assignCcCompatibleRequestDefaults,
   mergeCcCompatibleRequestDefaults,
